Added cancel buttons and username.
Also styling of those.

+++ listings/views/tickets.py
@@ -1,5 +1,6 @@
 from django.contrib.auth import get_user
 from django.contrib.auth.mixins import LoginRequiredMixin
+from django.http import HttpResponseRedirect
 from django.views.generic import DeleteView, CreateView, UpdateView
 from lazy import lazy_success_url
 from listings.forms import TicketEditForm
@@ -32,9 +33,12 @@ class EditTicketView(LoginRequiredMixin, UpdateView):
         context = super().get_context_data(**kwargs)
         ticket = Ticket.objects.filter(pk=int(self.kwargs["ticketid"])).first()
         context["ticket"] = ticket
+        context["back"] = self.get_success_url()
         return context
 
     def post(self, request, *args, **kwargs):
+        if "cancel" in request.POST:
+            return HttpResponseRedirect(self.get_success_url())
         return super().post(request, *args, **kwargs)
 
     def get_success_url(self) -> str:
@@ -48,6 +52,9 @@ class TicketDeleteView(LoginRequiredMixin, DeleteView):
     pk_url_kwarg = "ticketid"
 
     def form_valid(self, form):
+        if "cancel" in self.request.POST:
+            url = self.get_success_url()
+            return HttpResponseRedirect(url)
         image = self.object.image
         if not any(Ticket.objects.filter(image=image).exclude(pk=self.object.id)):
             image.delete()
